Add files via upload
This is a simple E-commerce prototype website created for a school project. It includes four pages: Home, Product, About, and Contact. The purpose of this project is to demonstrate basic website structure, an Ecommerce model, and an understanding of legal and security aspects in online shopping. No images or backend systems were used... this is a frontend learning project using only HTML.